+++ .sampo/changesets/issue-353-built-in-block-artifact-types.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split built-in block artifact type-source emitters into a focused helper module without changing generated scaffold output.
