+++ packages/wp-typia-project-tools/src/runtime/built-in-block-artifacts.ts
@@ -10,6 +10,13 @@ import type {
 import type {
 	BuiltInTemplateId,
 } from "./template-registry.js";
+import {
+	buildBasicTypesSource,
+	buildCompoundChildTypesSource,
+	buildCompoundTypesSource,
+	buildInteractivityTypesSource,
+	buildPersistenceTypesSource,
+} from "./built-in-block-artifact-types.js";
 
 const ALIGNMENT_VALUES = ["left", "center", "right"] as const;
 const BASIC_ALIGNMENT_VALUES = ["left", "center", "right", "justify"] as const;
@@ -110,24 +117,6 @@ interface CompoundChildAttributeVariables {
 	bodyPlaceholder: string;
 	childCssClassName?: string | null;
 	childTitle: string;
-}
-
-interface InterfaceMemberDefinition {
-	description?: AttributeDescription;
-	name: string;
-	optional?: boolean;
-	typeExpression: string;
-}
-
-interface InterfaceDefinition {
-	description?: AttributeDescription;
-	members: InterfaceMemberDefinition[];
-	name: string;
-}
-
-interface TypeAliasDefinition {
-	name: string;
-	value: string;
 }
 
 export interface BuiltInBlockArtifact {
@@ -381,70 +370,6 @@ function buildBlockJsonAttributes(
 
 function quote(value: string): string {
 	return JSON.stringify(value);
-}
-
-function emitDocComment(
-	description: AttributeDescription | undefined,
-	indent = "",
-): string[] {
-	if (!description) {
-		return [];
-	}
-
-	return [
-		`${indent}/**`,
-		...description.lines.map((line) =>
-			line.length === 0 ? `${indent} *` : `${indent} * ${line}`,
-		),
-		`${indent} */`,
-	];
-}
-
-function emitInterface(definition: InterfaceDefinition): string[] {
-	const lines = [
-		...emitDocComment(definition.description),
-		`export interface ${definition.name} {`,
-	];
-
-	for (const member of definition.members) {
-		lines.push(...emitDocComment(member.description, "\t"));
-		lines.push(
-			`\t${member.name}${member.optional ? "?" : ""}: ${member.typeExpression};`,
-		);
-	}
-
-	lines.push("}");
-	return lines;
-}
-
-function emitTypesModule({
-	preambleLines,
-	interfaces,
-	typeAliases,
-}: {
-	preambleLines: string[];
-	interfaces: InterfaceDefinition[];
-	typeAliases: TypeAliasDefinition[];
-}): string {
-	const sections: string[] = [];
-
-	if (preambleLines.length > 0) {
-		sections.push(preambleLines.join("\n"));
-	}
-
-	for (const definition of interfaces) {
-		sections.push(emitInterface(definition).join("\n"));
-	}
-
-	if (typeAliases.length > 0) {
-		sections.push(
-			typeAliases
-				.map((alias) => `export type ${alias.name} = ${alias.value};`)
-				.join("\n"),
-		);
-	}
-
-	return `${sections.join("\n\n")}\n`;
 }
 
 function stringifyBlockJsonDocument(document: Record<string, unknown>): string {
@@ -814,302 +739,6 @@ function buildCompoundChildAttributes(
 		bodyPlaceholder,
 		childCssClassName,
 		childTitle,
-	});
-}
-
-function buildBasicTypesSource(
-	variables: ScaffoldTemplateVariables,
-	attributes: readonly EmittedAttributeDefinition[],
-): string {
-	return emitTypesModule({
-		preambleLines: [
-			'import type { TextAlignment } from "@wp-typia/block-types/block-editor/alignment";',
-			"import type {",
-			"\tTypiaValidationError,",
-			"\tValidationResult,",
-			'} from "@wp-typia/block-runtime/validation";',
-			'import { tags } from "typia";',
-			"",
-			'export type { TypiaValidationError, ValidationResult } from "@wp-typia/block-runtime/validation";',
-		],
-		interfaces: [
-			{
-				description: {
-					lines: [
-						"Block attributes interface",
-						"Typia tags define runtime validation rules",
-					],
-				},
-				members: attributes.map((attribute) => ({
-					description: attribute.description,
-					name: attribute.name,
-					optional: attribute.optional,
-					typeExpression: attribute.typeExpression,
-				})),
-				name: `${variables.pascalCase}Attributes`,
-			},
-		],
-		typeAliases: [
-			{
-				name: `${variables.pascalCase}ValidationResult`,
-				value: `ValidationResult<${variables.pascalCase}Attributes>`,
-			},
-		],
-	});
-}
-
-function buildInteractivityTypesSource(
-	variables: ScaffoldTemplateVariables,
-	attributes: readonly EmittedAttributeDefinition[],
-): string {
-	return emitTypesModule({
-		preambleLines: [
-			'import type { TextAlignment } from "@wp-typia/block-types/block-editor/alignment";',
-			"import type {",
-			"\tTypiaValidationError,",
-			"\tValidationResult,",
-			'} from "@wp-typia/block-runtime/validation";',
-			'import { tags } from "typia";',
-			"",
-			'export type { TypiaValidationError, ValidationResult } from "@wp-typia/block-runtime/validation";',
-		],
-		interfaces: [
-			{
-				members: attributes.map((attribute) => ({
-					name: attribute.name,
-					optional: attribute.optional,
-					typeExpression: attribute.typeExpression,
-				})),
-				name: `${variables.pascalCase}Attributes`,
-			},
-			{
-				members: [
-					{
-						name: "clicks",
-						typeExpression: "number",
-					},
-					{
-						name: "isAnimating",
-						typeExpression: "boolean",
-					},
-					{
-						name: "isVisible",
-						typeExpression: "boolean",
-					},
-					{
-						name: "animation",
-						typeExpression: '"none" | "bounce" | "pulse" | "shake" | "flip"',
-					},
-					{
-						name: "maxClicks",
-						typeExpression: "number",
-					},
-				],
-				name: `${variables.pascalCase}Context`,
-			},
-		],
-		typeAliases: [
-			{
-				name: `${variables.pascalCase}ValidationResult`,
-				value: `ValidationResult<${variables.pascalCase}Attributes>`,
-			},
-		],
-	});
-}
-
-function buildPersistenceTypesSource(
-	variables: ScaffoldTemplateVariables,
-	attributes: readonly EmittedAttributeDefinition[],
-): string {
-	return emitTypesModule({
-		preambleLines: [
-			'import type { TextAlignment } from "@wp-typia/block-types/block-editor/alignment";',
-			"import type {",
-			"\tTypiaValidationError,",
-			"\tValidationResult,",
-			'} from "@wp-typia/block-runtime/validation";',
-			'import { tags } from "typia";',
-			"",
-			'export type { TypiaValidationError, ValidationResult } from "@wp-typia/block-runtime/validation";',
-		],
-		interfaces: [
-			{
-				members: attributes.map((attribute) => ({
-					name: attribute.name,
-					optional: attribute.optional,
-					typeExpression: attribute.typeExpression,
-				})),
-				name: `${variables.pascalCase}Attributes`,
-			},
-			{
-				members: [
-					{ name: "buttonLabel", typeExpression: "string" },
-					{ name: "bootstrapReady", typeExpression: "boolean" },
-					{ name: "canWrite", typeExpression: "boolean" },
-					{ name: "count", typeExpression: "number" },
-					{ name: "error", typeExpression: "string" },
-					{ name: "isBootstrapping", typeExpression: "boolean" },
-					{ name: "isLoading", typeExpression: "boolean" },
-					{ name: "isSaving", typeExpression: "boolean" },
-					{
-						name: "persistencePolicy",
-						typeExpression: '"authenticated" | "public"',
-					},
-					{ name: "postId", typeExpression: "number" },
-					{ name: "resourceKey", typeExpression: "string" },
-					{
-						name: "storage",
-						typeExpression: '"post-meta" | "custom-table"',
-					},
-					{ name: "isVisible", typeExpression: "boolean" },
-					{
-						name: "client",
-						optional: true,
-						typeExpression: `${variables.pascalCase}ClientState`,
-					},
-				],
-				name: `${variables.pascalCase}Context`,
-			},
-			{
-				members: [
-					{ name: "isHydrated", typeExpression: "boolean" },
-				],
-				name: `${variables.pascalCase}State`,
-			},
-			{
-				members: [
-					{ name: "bootstrapError", typeExpression: "string" },
-					{ name: "writeExpiry", typeExpression: "number" },
-					{ name: "writeNonce", typeExpression: "string" },
-					{ name: "writeToken", typeExpression: "string" },
-				],
-				name: `${variables.pascalCase}ClientState`,
-			},
-		],
-		typeAliases: [
-			{
-				name: `${variables.pascalCase}ValidationResult`,
-				value: `ValidationResult<${variables.pascalCase}Attributes>`,
-			},
-		],
-	});
-}
-
-function buildCompoundTypesSource(
-	variables: ScaffoldTemplateVariables,
-	attributes: readonly EmittedAttributeDefinition[],
-): string {
-	const persistenceEnabled = variables.compoundPersistenceEnabled === "true";
-
-	return emitTypesModule({
-		preambleLines: persistenceEnabled
-			? [
-				"import type {",
-				"\tTypiaValidationError,",
-				"\tValidationResult,",
-				'} from "@wp-typia/block-runtime/validation";',
-				'import { tags } from "typia";',
-				"",
-				'export type { TypiaValidationError, ValidationResult } from "@wp-typia/block-runtime/validation";',
-			]
-			: [
-				'import type { ValidationResult } from "@wp-typia/block-runtime/validation";',
-				'import { tags } from "typia";',
-				"",
-				'export type { ValidationResult } from "@wp-typia/block-runtime/validation";',
-			],
-		interfaces: [
-			{
-				members: attributes.map((attribute) => ({
-					name: attribute.name,
-					optional: attribute.optional,
-					typeExpression: attribute.typeExpression,
-				})),
-				name: `${variables.pascalCase}Attributes`,
-			},
-			...(persistenceEnabled
-				? [
-					{
-						members: [
-							{ name: "buttonLabel", typeExpression: "string" },
-							{ name: "bootstrapReady", typeExpression: "boolean" },
-							{ name: "canWrite", typeExpression: "boolean" },
-							{ name: "count", typeExpression: "number" },
-							{ name: "error", typeExpression: "string" },
-							{ name: "isBootstrapping", typeExpression: "boolean" },
-							{ name: "isLoading", typeExpression: "boolean" },
-							{ name: "isSaving", typeExpression: "boolean" },
-							{
-								name: "persistencePolicy",
-								typeExpression: '"authenticated" | "public"',
-							},
-							{ name: "postId", typeExpression: "number" },
-							{ name: "resourceKey", typeExpression: "string" },
-							{ name: "showCount", typeExpression: "boolean" },
-							{
-								name: "storage",
-								typeExpression: '"post-meta" | "custom-table"',
-							},
-							{
-								name: "client",
-								optional: true,
-								typeExpression: `${variables.pascalCase}ClientState`,
-							},
-						],
-						name: `${variables.pascalCase}Context`,
-					},
-					{
-						members: [{ name: "isHydrated", typeExpression: "boolean" }],
-						name: `${variables.pascalCase}State`,
-					},
-					{
-						members: [
-							{ name: "bootstrapError", typeExpression: "string" },
-							{ name: "writeExpiry", typeExpression: "number" },
-							{ name: "writeNonce", typeExpression: "string" },
-							{ name: "writeToken", typeExpression: "string" },
-						],
-						name: `${variables.pascalCase}ClientState`,
-					},
-				]
-				: []),
-		],
-		typeAliases: [
-			{
-				name: `${variables.pascalCase}ValidationResult`,
-				value: `ValidationResult<${variables.pascalCase}Attributes>`,
-			},
-		],
-	});
-}
-
-function buildCompoundChildTypesSource(
-	variables: ScaffoldTemplateVariables,
-	attributes: readonly EmittedAttributeDefinition[],
-): string {
-	return emitTypesModule({
-		preambleLines: [
-			'import type { ValidationResult } from "@wp-typia/block-runtime/validation";',
-			'import { tags } from "typia";',
-			"",
-			'export type { ValidationResult } from "@wp-typia/block-runtime/validation";',
-		],
-		interfaces: [
-			{
-				members: attributes.map((attribute) => ({
-					name: attribute.name,
-					optional: attribute.optional,
-					typeExpression: attribute.typeExpression,
-				})),
-				name: `${variables.pascalCase}ItemAttributes`,
-			},
-		],
-		typeAliases: [
-			{
-				name: `${variables.pascalCase}ItemValidationResult`,
-				value: `ValidationResult<${variables.pascalCase}ItemAttributes>`,
-			},
-		],
 	});
 }
